3. Application Using LCEL(LangChain Expression Language)/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
openai
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
pydantic==1.10.13